+++ Sources/RichTextKit/Attributes/RichTextAttributeReader+Colors.swift
@@ -9,39 +9,13 @@
 import Foundation
 
 public extension RichTextAttributeReader {
-
-    /// Get the background color at a certain range.
-    func richTextBackgroundColor(
-        at range: NSRange
+    
+    /// Get a certain rich text color at the provided range.
+    func richTextColor(
+        _ color: RichTextColor,
+        at range: NSRange? = nil
     ) -> ColorRepresentable? {
-        richTextAttribute(.backgroundColor, at: range)
-    }
-
-    /// Get the foreground color at a certain range.
-    func richTextForegroundColor(
-        at range: NSRange
-    ) -> ColorRepresentable? {
-        richTextAttribute(.foregroundColor, at: range)
-    }
-
-    /// Get the strikethrough color at a certain range.
-    func richTextStrikethroughColor(
-        at range: NSRange
-    ) -> ColorRepresentable? {
-        richTextAttribute(.strikethroughColor, at: range)
-    }
-
-    /// Get the stroke color at a certain range.
-    func richTextStrokeColor(
-        at range: NSRange
-    ) -> ColorRepresentable? {
-        richTextAttribute(.strokeColor, at: range)
-    }
-
-    /// Get the underline color at a certain range.
-    func richTextUnderlineColor(
-        at range: NSRange
-    ) -> ColorRepresentable? {
-        richTextAttribute(.underlineColor, at: range)
+        guard let attribute = color.attribute else { return nil }
+        return richTextAttribute(attribute, at: range)
     }
 }