+++ Sources/RichTextKit/Attributes/RichTextAttributeWriter+Colors.swift
@@ -10,16 +10,7 @@ import Foundation
 
 public extension RichTextAttributeWriter {
 
-    /// Get the value of a certain rich text color.
-    func richTextColor(
-        _ color: RichTextColor,
-        at range: NSRange? = nil
-    ) -> ColorRepresentable? {
-        guard let attribute = color.attribute else { return nil }
-        return richTextAttribute(attribute, at: range)
-    }
-
-    /// Set the current value of a certain rich text color.
+    /// Set a certain rich text color at the provided range.
     func setRichTextColor(
         _ color: RichTextColor,
         to val: ColorRepresentable,
@@ -28,13 +19,5 @@ public extension RichTextAttributeWriter {
         guard let attribute = color.attribute else { return }
         if richTextColor(color, at: range) == val { return }
         setRichTextAttribute(attribute, to: color, at: range)
-    }
-
-    /// Set the underline color at a certain range.
-    func setRichTextUnderlineColor(
-        _ color: ColorRepresentable,
-        at range: NSRange? = nil
-    ) {
-        setRichTextAttribute(.underlineColor, to: color, at: range)
     }
 }